scripts/pyproject.toml
# ============================================
# Self-Hosted AI Platform - Scripts Package
# ============================================
# Python 3.14+ with uv for dependency management

[project]
name = "self-hosted-ai-scripts"
version = "1.0.0"
description = "Infrastructure automation scripts for Self-Hosted AI Platform"
readme = "README.md"
license = { text = "MIT" }
requires-python = ">=3.12"
authors = [
    { name = "VectorWeight", email = "admin@vectorweight.com" }
]
keywords = ["kubernetes", "automation", "ai", "infrastructure"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Environment :: Console",
    "Intended Audience :: Developers",
    "Intended Audience :: System Administrators",
    "License :: OSI Approved :: MIT License",
    "Operating System :: POSIX :: Linux",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Programming Language :: Python :: 3.14",
    "Topic :: System :: Systems Administration",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]

dependencies = [
    # HTTP & API
    "httpx>=0.27.0",
    "aiofiles>=24.1.0",
    
    # Kubernetes
    "kubernetes>=31.0.0",
    "kubernetes-asyncio>=31.0.0",
    
    # CLI & UX
    "rich>=13.9.0",
    "typer>=0.12.0",
    "questionary>=2.0.0",
    
    # Data & Config
    "pyyaml>=6.0.0",
    "pydantic>=2.9.0",
    "pydantic-settings>=2.5.0",
    "python-dotenv>=1.0.0",
    
    # Security
    "cryptography>=43.0.0",
    "secrets-manager>=0.1.0",
    
    # Database
    "asyncpg>=0.30.0",
    "redis>=5.2.0",
    
    # Utilities
    "tenacity>=9.0.0",
    "arrow>=1.3.0",
    "jinja2>=3.1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.0",
    "pytest-asyncio>=0.24.0",
    "pytest-cov>=5.0.0",
    "ruff>=0.7.0",
    "mypy>=1.13.0",
    "pre-commit>=4.0.0",
]

[project.scripts]
shai = "cli.main:app"
shai-bootstrap = "bootstrap:main"
shai-validate = "validate_cluster:main"
shai-secrets = "secrets_manager:main"
shai-backup = "backup_restore:main"

[project.urls]
Homepage = "https://github.com/tzervas/self-hosted-ai"
Repository = "https://github.com/tzervas/self-hosted-ai"
Documentation = "https://github.com/tzervas/self-hosted-ai#readme"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["."]

# ============================================
# Tool Configuration
# ============================================

[tool.ruff]
target-version = "py312"
line-length = 100
exclude = [
    ".git",
    ".venv",
    "__pycache__",
    "build",
    "dist",
]

[tool.ruff.lint]
select = [
    "E",      # pycodestyle errors
    "W",      # pycodestyle warnings
    "F",      # Pyflakes
    "I",      # isort
    "B",      # flake8-bugbear
    "C4",     # flake8-comprehensions
    "UP",     # pyupgrade
    "ARG",    # flake8-unused-arguments
    "SIM",    # flake8-simplify
    "TCH",    # flake8-type-checking
    "PTH",    # flake8-use-pathlib
    "ERA",    # eradicate
    "PL",     # Pylint
    "RUF",    # Ruff-specific
]
ignore = [
    "PLR0913",  # Too many arguments
    "PLR2004",  # Magic values
]

[tool.ruff.lint.isort]
known-first-party = ["cli", "lib"]

[tool.mypy]
python_version = "3.12"
strict = true
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true

[[tool.mypy.overrides]]
module = [
    "kubernetes.*",
    "kubernetes_asyncio.*",
]
ignore_missing_imports = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
addopts = "-v --tb=short --cov=. --cov-report=term-missing"

[tool.coverage.run]
source = ["."]
omit = ["tests/*", "**/__pycache__/*"]
